+++ src/main/scala/ai/starlake/workflow/IngestionWorkflow.scala
@@ -26,23 +26,13 @@ import ai.starlake.job.infer.{InferSchema, InferSchemaConfig}
 import ai.starlake.job.ingest._
 import ai.starlake.job.load.LoadStrategy
 import ai.starlake.job.metrics.{MetricsConfig, MetricsJob}
-import ai.starlake.job.sink.bigquery.{
-  BigQueryJobBase,
-  BigQueryJobResult,
-  BigQueryLoadConfig,
-  BigQuerySparkJob
-}
+import ai.starlake.job.sink.bigquery.{BigQueryJobBase, BigQueryJobResult, BigQueryLoadConfig, BigQuerySparkJob}
 import ai.starlake.job.sink.es.{ESLoadConfig, ESLoadJob}
 import ai.starlake.job.sink.jdbc.{ConnectionLoadJob, JdbcConnectionLoadConfig}
 import ai.starlake.job.sink.kafka.{KafkaJob, KafkaJobConfig}
 import ai.starlake.job.transform.{AutoTask, TransformConfig}
 import ai.starlake.schema.generator.{Yml2DDLConfig, Yml2DDLJob}
-import ai.starlake.schema.handlers.{
-  LaunchHandler,
-  LocalStorageHandler,
-  SchemaHandler,
-  StorageHandler
-}
+import ai.starlake.schema.handlers.{LaunchHandler, LocalStorageHandler, SchemaHandler, StorageHandler}
 import ai.starlake.schema.model.Engine.BQ
 import ai.starlake.schema.model.Mode.{FILE, STREAM}
 import ai.starlake.schema.model._
@@ -50,6 +40,7 @@ import ai.starlake.utils._
 import better.files.File
 import com.google.cloud.bigquery.JobInfo.{CreateDisposition, WriteDisposition}
 import com.google.cloud.bigquery.{Schema => BQSchema}
+import com.typesafe.config.ConfigFactory
 import com.typesafe.scalalogging.StrictLogging
 import org.apache.hadoop.fs.Path
 import org.apache.spark.sql.catalyst.SQLConfHelper
@@ -652,11 +643,12 @@ class IngestionWorkflow(
   }
 
   def inferSchema(config: InferSchemaConfig): Try[File] = {
+    implicit val settings: Settings = Settings(ConfigFactory.load())
     val result = new InferSchema(
       config.domainName,
       config.schemaName,
       config.inputPath,
-      config.outputDir,
+      config.outputDir.getOrElse(DatasetArea.load.toString),
       config.withHeader,
       config.format
     ).run()